BLE api 수정 및 db형식에 맞게 변경

+++ src/main/java/devkor/com/teamcback/domain/ble/entity/BLEDevice.java
@@ -31,10 +31,10 @@ public class BLEDevice extends BaseEntity {
     private int capacity;
 
     @Column(name="default_count")
-    private int defaultCount;
+    private double defaultCount;
 
     @Column
-    private int ratio;
+    private double ratio;
 
     public BLEDevice(CreateBLEDeviceReq req, Place place) {
         this.deviceName = req.getDeviceName();
